api/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
huggingface-hub==0.18.0
pytest==7.4.0
coverage==7.3.2
